DataViews: Register the reset template and template part action like any third-party action
