+++ packages/editor/src/dataviews/actions/index.ts
@@ -7,6 +7,8 @@ import { type StoreDescriptor, dispatch } from '@wordpress/data';
  * Internal dependencies
  */
 import deletePost from './delete-post';
+import resetPost from './reset-post';
+
 // @ts-ignore
 import { store as editorStore } from '../../store';
 import { unlock } from '../../lock-unlock';
@@ -16,5 +18,6 @@ export default function registerDefaultActions() {
 		dispatch( editorStore as StoreDescriptor )
 	);
 
+	registerEntityAction( 'postType', '*', resetPost );
 	registerEntityAction( 'postType', '*', deletePost );
 }
