+++ packages/editor/src/dataviews/types.ts
@@ -11,11 +11,15 @@ export interface BasePost {
 	status?: PostStatus;
 	title: string | { rendered: string };
 	type: string;
+	id: string | number;
 }
 export interface TemplateOrTemplatePart extends BasePost {
-	type: 'template' | 'template-part';
+	type: 'wp_template' | 'wp_template_part';
 	source: string;
 	has_theme_file: boolean;
 }
 
 export type Post = TemplateOrTemplatePart | BasePost;
+
+// Will be unnecessary after typescript 5.0 upgrade.
+export type CoreDataError = { message?: string; code?: string };
